Change send page coin control section to a toggleable panel
This updates the coin control section on the send coins page to behave
like a toggleable panel. Clicking on the header opens the coin control
options, and clicking the label again hides them.

When the panel is closed, coin control does not apply to the selection
of inputs or the change address even when a user configured the fields
prior to toggling the feature. To emphasize this, a status label shows
when coin control settings influence a transaction.