Add empty placeholder to GUI recent transactions list
This adds a presentation nicety to the recent transactions list on the
overview page that shows an empty placeholder text and icon in wallets
with no transactions. It is a bit of polish that fills the empty space
that new users see with a fresh installation.

The placeholder is a widget that we can reuse in other components when
we need to display a "nothing here" or "no results" message.